Implement kubetest2 GCE Up() for cloud-provider-gcp
Implements the Up() method for the kubetest2 GCE deployer. This
version of Up() only supports the kube-up.sh in k/cloud-provider-gcp
for POC/MVP. Also sets path for kubeconfig to be in the kubetest2
artifacts directory. Chose to start with empty environment rather
than inheriting from the OS to guarantee functionality on minimal
information.