Default artifacts directory is an absolute path
Making this path absolute saves several headaches when calling
parts of the ginkgo tester. Specifically, if ginkgo does not
have an absolute path for the kubeconfig it will fail because
(most likely) it changes the working directory for some reason.
By modifying the artifacts directory where it is set initially,
all appends to it (like the GCE deployer does for its kubeconfig
path) are safe. Changing paths to absolute at the deployer level
could cause headaches if a user-specified artifacts directory
is an absolute path.