Default artifacts is absolute and ginkgo expands KUBECONFIG to abs
Odd behavior was encountered when running the ginkgo tester with
non-absolute paths for ARTIFACTS after running the GCE deployer.
Observed behavior shows that when the ginkgo executable is called
with a non-absolute path for KUBECONFIG it errors. Because the GCE
deployer merely adds paths on top of ARTIFACTS, it was returning a
KUBECONFIG (sent to the tester) of ARTIFACTS/kubetest2-kubeconfig
which was a relative path because ARTIFACTS is, by default,
"_artifacts".

This commit addresses the problem at 2 levels: it changes the
default artifacts dir to have an absolute path and then ensures an
absolute path for the kubeconfig passed to the ginkgo executable.
While this problem could be addressed by only ensuring the
kubeconfig path is absolute in the ginkgo tester, it is also
better to have a good default for artifacts in case of future
problems.