tester-ginkgo: add klog flags
There are some klog.V(2) logs in the tester that were never getting
displayed because klog flags weren't exposed to allow a user to set v=2
for the tester.

Fix that by adding klog flags to the ginkgo tester

e.g. the following will set v=1 for the noop deployer, v=2 for the
ginkgo tester

    kubetest2 noop --v=1 --test=ginkgo -- --v=2