Remove deprecated --kube-root kind argument
The --kube-root flag for kind has been deprecated and will be removed
soon. The current method for the equivalent functionality is to pass
this path to the k8s source directory as the first non-flag argument to
"kind build node-image".

This removes the deprecated usage and updates the help text to reflect
the new implementation.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>